+++ src/main/java/inu/codin/codin/domain/lecture/repository/elasticsearch/LectureElasticRepository.java
@@ -1,6 +1,6 @@
 package inu.codin.codin.domain.lecture.repository.elasticsearch;
 
-import inu.codin.codin.domain.lecture.document.LectureDocument;
+import inu.codin.codin.domain.lecture.search.document.LectureDocument;
 import inu.codin.codin.domain.lecture.entity.SortingOption;
 import inu.codin.codin.global.common.entity.Department;
 import org.springframework.data.domain.Page;
