+++ src/main/java/inu/codin/codin/domain/lecture/service/LectureIndexService.java
@@ -1,7 +1,7 @@
 package inu.codin.codin.domain.lecture.service;
 
 import inu.codin.codin.domain.lecture.converter.LectureToDocumentConverter;
-import inu.codin.codin.domain.lecture.document.LectureDocument;
+import inu.codin.codin.domain.lecture.search.document.LectureDocument;
 import inu.codin.codin.domain.lecture.event.LectureDeletedEvent;
 import inu.codin.codin.domain.lecture.event.LectureSavedEvent;
 import inu.codin.codin.domain.lecture.repository.elasticsearch.LectureElasticRepository;
